requirements.txt
fastapi==0.135.1
google_search_results==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.32.5
uvicorn